+++ R/prepare_data_dot.R
@@ -22,6 +22,9 @@
 #' @param abort_msg2 An additional message to append to the error message in
 #'   case data-dot-injection is not permitted (when
 #'   `.SciViews.implicit.data.dot != TRUE`, see example).
+#' @param abort_msg3 An error message when `.` is not found.
+#' @param abort_msg4 An additional explanation when `.`is not found.
+#' @param abort_msg5 A hint to read the documentation of the data-dot mechanism.
 #' @param abort_frame The environment to use for the error message, by default,
 #'   the caller environment (should rarely be changed).
 #'
@@ -37,37 +40,47 @@
 #' `FALSE` otherwise. The result from evaluating the data-dot-injected call for
 #' `recall_with_data_dot()`.
 #' @export
+#' @seealso [data_dot_mechanism]
 #'
 #' @examples
+#' # Use this (rename) function to get extra-info in the error message about the
+#' # data-dot mechanism automatically
+#' stop <- stop_
+#'
 #' # Here is how you create a data-dot function
 #' my_subset <- function(.data = (.), i, j) {
-#'   # This makes it a data-dot function
 #'   if (!prepare_data_dot(.data))
 #'     return(recall_with_data_dot())
 #'
-#'   # Code of the function
-#'   # Second argument (i here) must not be a data.frame to avoid confusion
-#'   message(".env has ", paste(names(.env), collapse = ", "))
+#'   if (!is.numeric(i))
+#'     stop_("{.arg i} must be numeric") # Use this function
+#'   if (i == 1)
+#'     message(".env has ", paste(names(.env), collapse = ", "))
 #'   .data[i, j]
 #' }
-#' dtf1 <- data.frame(x = 1:3, y = 4:6)
-#' my_subset(dtf1, 1, 'y')
-#' # If .data is in '.', it can be omitted
-#' .= dtf1
-#' my_subset(1, 'y')
 #'
-#' # This mechanism is potentially confusing. You can inactivate it anywhere:
+#' dtf1 <- data.frame(x = 1:3, y = 4:6)
+#' # The message shows the objects available in the function environment
+#' my_subset(dtf1, 1, 'y')
+#' # This is wrong
+#'try(my_subset(dtf1, 'y'))
+#' .= dtf1
+#' my_subset(2, 'y')
+#' # Error message with indication that the data-dot mechanism is activated
+#' try(my_subset('y'))
+#' # Data-dot mechanism not activate, but dot object used
+#' try(my_subset(., 'y'))
+#' # Wrong .data=
+#' try(my_subset(.data = 'y'))
+#'
+#' # When data-dot is not permitted...
 #' .SciViews.implicit.data.dot <- FALSE
-#' # This time next call is wrong
-#' try(my_subset(1, 'y'))
-#' # You must indicate '.' explicitly in that case:
-#' my_subset(., 1, 'y')
-#' rm(.SciViews.implicit.data.dot) # Reactivate it
-#' my_subset(1, 'y') # Implicit again
-#' # Note that, if you have not defined '.' and try to use it, you got
-#' # an error:
+#' try(my_subset(2, 'y'))
+#' rm(.SciViews.implicit.data.dot)
+#'
+#' # When `.` is not found...
 #' rm(.)
-#' try(my_subset(1, 'y'))
+#' try(my_subset(2, 'y'))
 prepare_data_dot <- function(x, is_top_call = TRUE) {
   if (!missing(x) && inherits(x, 'data.frame')) {
     pf <- parent.frame()
@@ -106,17 +119,20 @@ prepare_data_dot2 <- function(x, y, is_top_call = TRUE) {
 
 #' @rdname prepare_data_dot
 #' @export
-recall_with_data_dot <- function(call, arg = '.data', value = as.symbol('.'),
+recall_with_data_dot <- function(call, arg = '.data', value = quote((.)),
   env = parent.frame(2L),
   abort_msg = gettextf("`%s` must be a 'data.frame'.", arg),
   abort_msg2 = gettext("Implicit data-dot (.) not permitted"),
+  abort_msg3 = gettext("Data-dot mechanism activated, but no `.` object found."),
+  abort_msg4 = gettextf("Define `.` before calling this function, or provide `%s=` explicitly.", arg),
+  abort_msg5 = gettextf("See {.help svMisc::data_dot_mechanism} for more infos."),
   abort_frame = parent.frame()) {
 
   if (missing(call))
     call <- sys.call(-1L)
 
   if (isFALSE(.SciViews.implicit.data.dot))
-    abort(c(abort_msg, i = abort_msg2), .frame = abort_frame)
+    abort(c(abort_msg, x = abort_msg2), .frame = abort_frame)
 
   if (any(names(call) == arg))
     abort(abort_msg, .frame = abort_frame)
@@ -131,22 +147,35 @@ recall_with_data_dot <- function(call, arg = '.data', value = as.symbol('.'),
   call[[2]] <- value
   names(call)[2] <- arg
 
-  eval_bare(call, env)
+  # Check that "." exists
+  if (!exists('.')) {
+    cli_abort(c(abort_msg3, i = abort_msg4, i = abort_msg5),
+      .frame = abort_frame)
+  }
+
+  # Indicate that the data-dot mechanisms was triggered
+  pf <- parent.frame()
+  pf$._data_dot_. <- TRUE
+
+  Exec(call, env)
 }
 
 #' @rdname prepare_data_dot
 #' @export
 recall_with_data_dot2 <- function(call, arg = 'x', arg2 = 'y',
-  value = as.symbol('.'), env = parent.frame(2L),
+  value = quote((.)), env = parent.frame(2L),
   abort_msg = gettextf("`%s` and `%s` must both be 'data.frame'.", arg, arg2),
   abort_msg2 = gettext("Implicit data-dot (.) not permitted"),
+  abort_msg3 = gettext("Data-dot mechanism activated, but no `.` object found."),
+  abort_msg4 = gettextf("Define `.` before calling this function, or provide `%s=` explicitly.", arg),
+  abort_msg5 = gettextf("See {.help svMisc::data_dot_mechanism} for more infos."),
   abort_frame = parent.frame()) {
 
   if (missing(call))
     call <- sys.call(-1L)
 
   if (isFALSE(.SciViews.implicit.data.dot))
-    abort(c(abort_msg, i = abort_msg2), .frame = abort_frame)
+    abort(c(abort_msg, x = abort_msg2), .frame = abort_frame)
 
   if (any(names(call) == arg))
     abort(abort_msg, .frame = abort_frame)
@@ -161,5 +190,58 @@ recall_with_data_dot2 <- function(call, arg = 'x', arg2 = 'y',
   call[[2]] <- value
   names(call)[2] <- arg
 
-  eval_bare(call, env)
+  # Check that "." exists
+  if (!exists('.')) {
+    cli_abort(c(abort_msg3, x = abort_msg4, i = abort_msg5),
+      .frame = abort_frame)
+  }
+
+  Exec(call, env)
 }
+
+#' The data-dot mechanism
+#'
+#' @name data_dot_mechanism
+#' @description The **data-dot** mechanism injects automatically `.data = .` in
+#' the call to a function when it detects it is necessary (most of the time,
+#' when `.data=` is missing, or a unnamed first argument is not suitable as
+#' `.data`, i.e., it is not a **data.frame**).
+#' This is useful to avoid having to avoid writing `.` "everywhere" in your
+#' functions when you use the explicit pipe operator `%>.%`, or with `.= ...`
+#' constructs.
+#' The data-dot mechanism may fail with an error message if it cannot inject
+#' `.` as `.data=`, or when `.` is not found. It may also be prohibited if the
+#' variable `.SciViews.implicit.data.dot` is set to
+#' ,0 `FALSE` (see examples).
+#' @seealso [prepare_data_dot()]
+#' @examples
+#' # Here is how you create a data-dot function
+#' my_subset <- function(.data = (.), i, j) {
+#'   # This makes it a data-dot function
+#'   if (!prepare_data_dot(.data))
+#'     return(recall_with_data_dot())
+#'
+#'   # Code of the function
+#'   # Second argument (i here) must not be a data.frame to avoid confusion
+#'   message(".env has ", paste(names(.env), collapse = ", "))
+#'   .data[i, j]
+#' }
+#' dtf1 <- data.frame(x = 1:3, y = 4:6)
+#' my_subset(dtf1, 1, 'y')
+#' # If .data is in '.', it can be omitted
+#' .= dtf1
+#' my_subset(1, 'y')
+#'
+#' # This mechanism is potentially confusing. You can inactivate it anywhere:
+#' .SciViews.implicit.data.dot <- FALSE
+#' # This time next call is wrong
+#' try(my_subset(1, 'y'))
+#' # You must indicate '.' explicitly in that case:
+#' my_subset(., 1, 'y')
+#' rm(.SciViews.implicit.data.dot) # Reactivate it
+#' my_subset(1, 'y') # Implicit again
+#' # Note that, if you have not defined '.' and try to use it, you got
+#' # an error:
+#' rm(.)
+#' try(my_subset(1, 'y'))
+NULL