+++ src/exprfuns.c
@@ -1,4 +1,3 @@
-#define R_NO_REMAP
 #include <R.h>
 #include <Rinternals.h>