+++ vignettes/svMisc.R
@@ -34,23 +34,6 @@ for (i in 0:75) {
 }
 
 ## -----------------------------------------------------------------------------
-foo <- structure(function(x, type = c("histogram", "boxplot"), ...) {
-  type <- match.arg(type, c("histogram", "boxplot"))
-  switch(type,
-    histogram = hist(x, ...),
-    boxplot = boxplot(x, ...),
-    stop("unknow type")
-  )
-}, class = c("function", "subsettable_type"))
-foo
-
-# This function can be used as usual:
-foo(rnorm(50), type = "histogram")
-# ... but also this way:
-foo$histogram(rnorm(50))
-foo$boxplot(rnorm(50))
-
-## -----------------------------------------------------------------------------
 captured <- capture_all(parse_text('1:2 + 1:3'), split = FALSE)
 captured