+++ NEWS.md
@@ -1,6 +1,8 @@
 # svMisc 1.5.2
 
--   The `expr_funs()` function is added to return all functions used in an expression or a call (it uses c code adapted fro R base).
+-   `retarget()` added (change easily the environment of a formula).
+
+-   `expr_funs()` is added to return all functions used in an expression or a call (it uses c code adapted fro R base).
 
 # svMisc 1.5.1