+++ NEWS.md
@@ -1,5 +1,9 @@
 # svMisc 1.5.2
 
+-   `stop_()`, `warning_()` and `stop_top_call()` added.
+
+-   `?` is renamed `%?%` to avoid conflict with `utils::?`.
+
 -   `gettext`, `gettextf` and `ngettext` renamed `gettext_`, `gettexf_` and `ngettext_` to avoid forcing using them. One must rename them in a package that uses the language switch feature of these function.
 
 -   `retarget()` added (change easily the environment of a formula).